% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{bench_outer_product}
\alias{bench_outer_product}
\title{Outer Product}
\usage{
bench_outer_product(x, y)
}
\arguments{
\item{x}{1st data vector}

\item{y}{2nd data vector}
}
\description{
Computes outer product of two vectors: result[i,j] = x[i] * y[j].
 Tests proper matrix allocation and filling patterns.
 This is a case where you must allocate the full result upfront.
}
