% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{bench_prealloc_weighted_sum}
\alias{bench_prealloc_weighted_sum}
\title{Pre-allocated Vector Sum}
\usage{
bench_prealloc_weighted_sum(x, y, w)
}
\arguments{
\item{x}{1st numeric vector}

\item{y}{2nd numeric vector}

\item{w}{weight vector}
}
\description{
Computes weighted sum of two vectors with pre-allocation.
 Test goal: create result once, no growing vectors, no unnecessary allocations.
}
