% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcppbenchmark-package.R
\docType{package}
\name{Rcppbenchmark-package}
\alias{Rcppbenchmark}
\alias{Rcppbenchmark-package}
\title{Rcppbenchmark: Checking speed improvement/worsening with different C++ standards for Rcpp}
\description{
A benchmark package to test the speed of various C++ implementations using the Rcpp package. The benchmarks include operations such as cumulative sums, outer products, pairwise distances, grouped means, and bootstrap resampling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://WEBSITE.COM}
  \item Report bugs at \url{https://github.com/USERNAME/PKGNAME/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [copyright holder]

}
\keyword{internal}
