% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcppbenchmark-package.R
\docType{package}
\name{Rcppbenchmark-package}
\alias{Rcppbenchmark}
\alias{Rcppbenchmark-package}
\title{Rcppbenchmark: Checking speed improvement/worsening with different C++ standards for Rcpp}
\description{
Benchmark adapted from the official Armadillo documentation to test the speed of different C++ standards when using the cpp4r package. The idea is to check how compiling with GCC or Clang using C++11, C++14, C++17, C++20 or C++23 affects the speed of various linear algebra operations.
}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [copyright holder]

}
\keyword{internal}
